Add test for symlink output
